requirements.txt
pkg-resources==0.0.0
python-slugify==4.0.1
selenium==3.141.0
text-unidecode==1.3
urllib3==1.26.2
